Document DataFrame indexing features and remove redundant cell access tests
