+++ docs/StardustDocs/topics/_shadow_resources.md
@@ -154,6 +154,8 @@
 <resource src="valuesOnGroupByAllColumns_properties.html"></resource>
 <resource src="valuesOnGroupByRenameColumns_properties.html"></resource>
 <resource src="valuesOnGroupBySelectedColumns_properties.html"></resource>
+<resource src="indexingDf.html"></resource>
+<resource src="indicesGetRows.html"></resource>
 <resource src="dfLeftImplicit.html"></resource>
 <resource src="dfRightImplicit.html"></resource>
 <resource src="notebook_test_join_10.html"></resource>
