+++ docs/StardustDocs/topics/access.md
@@ -16,7 +16,7 @@ df.values() // Sequence<Any?>
 <!---END-->
 
 **Learn how to:**
-* [Access data by index](indexing.md)
+* [Access a cell by index](indexing.md#access-a-cell-by-index)
 * [Get row indices](indexing.md#row-indices)
 * [Iterate over data](iterate.md)
 * [Get a single row](getRow.md)
