+++ docs/StardustDocs/topics/indexing.md
@@ -1,6 +1,19 @@
 [//]: # (title: Indexing)
 
-<!---IMPORT org.jetbrains.kotlinx.dataframe.samples.api.Access-->
+<!---IMPORT org.jetbrains.kotlinx.dataframe.samples.api.IndexingSamples-->
+
+The examples on this page use the following [`DataFrame`](DataFrame.md):
+
+<!---FUN indexingDf-->
+
+```kotlin
+df
+```
+
+<!---END-->
+<inline-frame src="./resources/api/indexing/indexingDf.html" width="100%" height="500px"></inline-frame>
+
+## Access a cell by index
 
 <!---FUN getCell-->
 <tabs>
@@ -24,11 +37,8 @@ df[1]["age"]
 
 ## Row indices
 
-<!---IMPORT org.jetbrains.kotlinx.dataframe.samples.api.IndicesSamples-->
-
 `indices()` returns the indices of all rows of the [`DataFrame`](DataFrame.md) as an `IntRange` —
-from `0` to the number of rows minus one.
-The examples below use the same ten-row dataframe as [`count`](count.md).
+from `0` to the number of rows minus one:
 
 <!---FUN indices-->
 
@@ -46,20 +56,29 @@ as a `List<Int>` in row order:
 <tab title="Properties">
 
 ```kotlin
-df.indices { city == "Moscow" } // [2, 6]
+val moscow = df.indices { city == "Moscow" } // [2, 6]
 ```
 
 </tab>
 <tab title="Strings">
 
 ```kotlin
-df.indices { "city"<String?>() == "Moscow" } // [2, 6]
+val moscow = df.indices { "city"<String?>() == "Moscow" } // [2, 6]
 ```
 
 </tab></tabs>
 <!---END-->
 
-Use the result to [slice those rows](sliceRows.md) out of the dataframe.
+Use the result to [slice those rows](sliceRows.md) out of the dataframe:
 
-See also [`rowsCount`](rowsCount.md) and [`filter`](filter.md),
-which returns the matching rows instead of their indices.
+<!---FUN indicesGetRows-->
+
+```kotlin
+df.getRows(df.indices { city == "Moscow" })
+```
+
+<!---END-->
+<inline-frame src="./resources/api/indexing/indicesGetRows.html" width="100%" height="500px"></inline-frame>
+
+See also [`rowsCount`](rowsCount.md), which returns how many rows there are rather than their indices,
+and [`filter`](filter.md), which returns the matching rows themselves.
